Syncing up Model branch for View Additions

+++ src/Model/Location.java
@@ -25,11 +25,11 @@ public class Location {
 		isCurrent=false;
 		visited=false;
 	}
-	public double getLat_dd(){
-		return coord.dd_lat;
+	public int getLat_dd(){
+		return (int)coord.dd_lat;
 	}
-	public double getLon_dd(){
-		return coord.dd_long;
+	public int getLon_dd(){
+		return (int)coord.dd_long;
 	}
 	public String getId() {
 		return id;
